Fully align compression and decompression
