+++ README.md
@@ -19,7 +19,7 @@ These are the middlewares included in this crate:
 
 - `AddExtension`: Stick some shareable value in [request extensions].
 - `Compression`: Compression response bodies.
-- `Decompression`: Decompress request bodies.
+- `Decompression`: Decompress response bodies.
 
 Middlewares uses the [`http`] crate as the HTTP interface so they're compatible with any library or framework that also uses [`http`]. For example hyper and actix.
