+++ tower-http/src/compression/layer.rs
@@ -1,0 +1,76 @@
+use super::Compression;
+use crate::accept_encoding::AcceptEncoding;
+use tower_layer::Layer;
+
+/// Compress response bodies of the underlying service.
+///
+/// This uses the `Accept-Encoding` header to pick an appropriate encoding and adds the
+/// `Content-Encoding` header to responses.
+#[derive(Clone, Debug, Default)]
+pub struct CompressionLayer {
+    _priv: (),
+    accept: AcceptEncoding,
+}
+
+impl<S> Layer<S> for CompressionLayer {
+    type Service = Compression<S>;
+
+    fn layer(&self, inner: S) -> Self::Service {
+        Compression::new(inner)
+    }
+}
+
+impl CompressionLayer {
+    /// Create a new [`CompressionLayer`]
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Sets whether to enable the gzip encoding.
+    #[cfg(feature = "compression-gzip")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "compression-gzip")))]
+    pub fn gzip(self, enable: bool) -> Self {
+        self.accept.set_gzip(enable);
+        self
+    }
+
+    /// Sets whether to enable the Deflate encoding.
+    #[cfg(feature = "compression-deflate")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "compression-deflate")))]
+    pub fn deflate(self, enable: bool) -> Self {
+        self.accept.set_deflate(enable);
+        self
+    }
+
+    /// Sets whether to enable the Brotli encoding.
+    #[cfg(feature = "compression-br")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "compression-br")))]
+    pub fn br(self, enable: bool) -> Self {
+        self.accept.set_br(enable);
+        self
+    }
+
+    /// Disables the gzip encoding.
+    ///
+    /// This method is available even if the `gzip` crate feature is disabled.
+    pub fn no_gzip(self) -> Self {
+        self.accept.set_gzip(false);
+        self
+    }
+
+    /// Disables the Deflate encoding.
+    ///
+    /// This method is available even if the `deflate` crate feature is disabled.
+    pub fn no_deflate(self) -> Self {
+        self.accept.set_deflate(false);
+        self
+    }
+
+    /// Disables the Brotli encoding.
+    ///
+    /// This method is available even if the `br` crate feature is disabled.
+    pub fn no_br(self) -> Self {
+        self.accept.set_br(false);
+        self
+    }
+}
