+++ tower-http/src/decompression/mod.rs
@@ -1,0 +1,13 @@
+//! Middleware that decompresses response bodies.
+
+mod body;
+mod future;
+mod layer;
+mod service;
+
+pub use self::{
+    body::{DecompressionBody, Error},
+    future::ResponseFuture,
+    layer::DecompressionLayer,
+    service::Decompression,
+};
